Tidy up for Renne
