+++ spidermon/contrib/validation/jsonschema/tools.py
@@ -3,16 +3,19 @@ import logging
 import json
 from scrapy.utils.misc import load_object
 import six
-from spidermon.utils import web
+from six.moves.urllib.parse import urlparse
+from six.moves.urllib.request import urlopen
+
+logger = logging.getLogger(__name__)
 
 
 def get_schema_from(source):
-    if web.is_schema_url(source):
-        schema = web.get_contents(source)
+    if is_schema_url(source):
+        schema = get_contents(source)
         try:
             return json.loads(schema)
         except Exception as e:
-            logging.exception(
+            logger.exception(
                 str(e) + "\nCould not parse schema from '{}'".format(source)
             )
     elif source.endswith(".json"):
@@ -20,7 +23,7 @@ def get_schema_from(source):
             try:
                 return json.load(f)
             except Exception as e:
-                logging.exception(
+                logger.exception(
                     str(e) + "\nCould not parse schema in '{}'".format(source)
                 )
     else:
@@ -28,3 +31,19 @@ def get_schema_from(source):
         if isinstance(schema, six.string_types):
             return json.loads(schema)
         return schema
+
+
+def is_schema_url(path):
+    result = urlparse(path)
+    try:
+        return all([result.scheme, result.netloc, result.path])
+    except AttributeError:
+        return False
+
+
+def get_contents(url):
+    try:
+        with urlopen(url) as f:
+            return f.read().decode("utf-8")
+    except Exception as e:
+        logger.exception(str(e) + "\nFailed to get '{}'".format(url))
